Added pin definitions from schmatic

+++ software/src/types/maze_state.c
@@ -3,6 +3,7 @@
  */
 
 
+#include "stdint.h"
 #include "maze_state.h"
 
 /* initialize the state of the maze
@@ -10,11 +11,11 @@
  * wall in such a way that there are no duplicates.
  */
 void init_maze_state(maze_state_t* maze) {
-    for (int i = 0; i < sizeof(maze->wall_buffer) / sizeof(maze->wall_buffer[0]); ++i) {
+    for (uint16_t i = 0; i < sizeof(maze->wall_buffer) / sizeof(maze->wall_buffer[0]); ++i) {
         maze->wall_buffer[i].exists = 0.5;
     }
-    for (int x = 0; x < MAX_MAZE_WIDTH; ++x) {
-        for (int y = 0; y < MAX_MAZE_HEIGHT; ++y) {
+    for (uint16_t x = 0; x < MAX_MAZE_WIDTH; ++x) {
+        for (uint16_t y = 0; y < MAX_MAZE_HEIGHT; ++y) {
             
             maze->cells[x][y].north = &maze->wall_buffer[x * (MAX_MAZE_WIDTH + 1) + (y + 1)];
             maze->cells[x][y].east = &maze->wall_buffer[MAX_MAZE_HEIGHT * (MAX_MAZE_WIDTH + 1) + (x + 1) * MAX_MAZE_WIDTH + y];
